Add bootstrap-safe runtime entrypoint
